# activiti-cloud-query-service/activiti-cloud-starter-query/src/test/resources/swagger-expected.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
...
    "/v1/process-instances/{processInstanceId}/subprocesses": {
      "get": {
        "tags": [
          "process-instance-controller"
        ],
        "summary": "subprocesses",
        "operationId": "subprocesses",
        "parameters": [
          {
            "name": "processInstanceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "maxItems",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "skipCount",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "initiator",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "processDefinitionName",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "lastModifiedTo",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "serviceTasks",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "processDefinitionKey",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startTo",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "completedTo",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "rootProcessInstanceId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tasks",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "lastModifiedFrom",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "processDefinitionId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "variables",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sequenceFlows",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startFrom",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "completedFrom",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "completedDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "parentId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "suspendedFrom",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "activities",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "businessKey",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "lastModified",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "suspendedDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "startDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "processDefinitionVersion",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "CREATED",
                "RUNNING",
                "SUSPENDED",
                "CANCELLED",
                "COMPLETED"
              ]
            }
          },
          {
            "name": "suspendedTo",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntryResponseContentActivitiErrorMessage"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntryResponseContentActivitiErrorMessage"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntryResponseContentActivitiErrorMessage"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponseContentQueryCloudProcessInstance_General"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponseContentQueryCloudProcessInstance_General"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "oauth": []
          }
        ]
      }
    },
    "/v1/process-instances/{processInstanceId}/diagram": {
      "get": {
        "tags": [
          "process-instance-diagram-controller"
        ],
        "summary": "getProcessDiagram",
        "operationId": "getProcessDiagram",
        "parameters": [
          {
            "name": "processInstanceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntryResponseContentActivitiErrorMessage"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntryResponseContentActivitiErrorMessage"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntryResponseContentActivitiErrorMessage"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "image/svg+xml": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "oauth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActivitiErrorMessage": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "code": {
            "type": "integer",
            "format": "int32"
          }
        },
        "title": "ActivitiErrorMessage"
      },
      "EntryResponseContentActivitiErrorMessage": {
        "type": "object",
        "properties": {
          "entry": {
            "$ref": "#/components/schemas/ActivitiErrorMessage"
          }
        },
        "title": "EntryResponseContentActivitiErrorMessage"
      },
...
      "EntryResponseContentQueryCloudProcessInstance_General": {
        "type": "object",
        "properties": {
          "entry": {
            "$ref": "#/components/schemas/QueryCloudProcessInstance_General"
          }
        },
        "title": "EntryResponseContentQueryCloudProcessInstance_General"
      },
      "QueryCloudProcessInstance_General": {
        "type": "object",
        "properties": {
          "subprocesses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QueryCloudSubprocessInstance_General"
            },
            "uniqueItems": true
          },
          "appName": {
            "type": "string"
          },
          "serviceName": {
            "type": "string"
          },
          "serviceType": {
            "type": "string"
          },
          "serviceFullName": {
            "type": "string"
          },
          "serviceVersion": {
            "type": "string"
          },
          "appVersion": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "processDefinitionVersion": {
            "type": "integer",
            "format": "int32"
          },
          "processDefinitionKey": {
            "type": "string"
          },
          "processDefinitionName": {
            "type": "string"
          },
          "rootProcessInstanceId": {
            "type": "string"
          },
          "parentId": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "CREATED",
              "RUNNING",
              "SUSPENDED",
              "CANCELLED",
              "COMPLETED"
            ]
          },
          "startDate": {
            "type": "string",
            "format": "date-time"
          },
          "businessKey": {
            "type": "string"
          },
          "initiator": {
            "type": "string"
          },
          "processDefinitionId": {
            "type": "string"
          },
          "completedDate": {
            "type": "string",
            "format": "date-time"
          }
        },
        "title": "QueryCloudProcessInstance_General"
      },
      "QueryCloudSubprocessInstance_General": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "processDefinitionName": {
            "type": "string"
          }
        },
        "title": "QueryCloudSubprocessInstance_General"
      },
...
      "PaginationMetadata_General": {
        "type": "object",
        "properties": {
          "skipCount": {
            "type": "integer",
            "format": "int64"
          },
          "maxItems": {
            "type": "integer",
            "format": "int64"
          },
          "count": {
            "type": "integer",
            "format": "int64"
          },
          "hasMoreItems": {
            "type": "boolean"
          },
          "totalItems": {
            "type": "integer",
            "format": "int64"
          }
        },
        "title": "PaginationMetadata_General"
      },
      "EntriesResponseContentQueryCloudProcessInstance_General": {
        "type": "object",
        "properties": {
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EntryResponseContentQueryCloudProcessInstance_General"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationMetadata_General"
          }
        },
        "title": "EntriesResponseContentQueryCloudProcessInstance_General"
      },
      "ListResponseContentQueryCloudProcessInstance_General": {
        "type": "object",
        "properties": {
          "list": {
            "$ref": "#/components/schemas/EntriesResponseContentQueryCloudProcessInstance_General"
          }
        },
        "title": "ListResponseContentQueryCloudProcessInstance_General"
      },
...
    }
  },
...
}